refactor: extract getCharacterDetails function
