+++ src/lib/character/getCharacterId.ts
@@ -1,4 +1,4 @@
-import { Character } from "../types/character";
+import { Character } from "../../types/character";
 
 export default function getCharacterId(character: Character) {
   return character.url.split("/").filter(Boolean).pop();
